flashearV2.sh: Added comments and check if exists \'pv\'

+++ flashearV2.sh
@@ -24,6 +24,13 @@ resetColor(){
 
 currentUser=`whoami`
 
+####################
+####     pv    #####
+####################
+
+EXISTSPV=`command -v pv`
+
+# Clear screen
 clear
 
 if [ $currentUser == "root" ]; then
@@ -49,12 +56,15 @@ if [ $currentUser == "root" ]; then
 	read -p "Choose the image: " flash
 
 	if [ $flash -gt ${#imageArray[@]} ] || [ $flash -lt 0 ]; then
+		# Bad number
 		clear
 		redColor
 		echo "Invalid number, launch the script again and choose one number of the list"
 		resetColor
+		exit 1
 
 	else
+		# Clear screen
 		clear
 		flash=$(($flash-1))
 		# Check file extension 
@@ -62,16 +72,35 @@ if [ $currentUser == "root" ]; then
 		if [ -z $EXTENSION ]; then
 			EXTENSION=`echo $$IMAGE_PATH/${imageArray[$flash]} | cut -d "." -f2`
 			case $EXTENSION in
-				"img" ) if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
-							echo "Unmounted correctly"
-							echo "Flashing, please wait..."
-							SIZE=`du -h $IMAGE_PATH/${imageArray[$flash]} | cut -d "," -f1`G
-							dd if=$IMAGE_PATH/${imageArray[$flash]} | pv -s $SIZE | dd of=$FLASH_DEVICE bs=8m
+				"img" ) # Check if 'pv' command exists
+						if [ -z $EXISTSPV ]; then
+							if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
+								# Flash with progress bar
+								echo "Unmounted correctly"
+								echo "Flashing, please wait..."
+								SIZE=`du -h $IMAGE_PATH/${imageArray[$flash]} | cut -d "," -f1`G
+								dd if=$IMAGE_PATH/${imageArray[$flash]} | pv -s $SIZE | dd of=$FLASH_DEVICE bs=8m
+							else
+								clear
+								redColor
+								echo "Please, change the device number"
+								resetColor
+								exit 1
+							fi
 						else
-							clear
-							redColor
-							echo "Please, change the device number"
-							resetColor
+							# Flash without progress bar
+							if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
+								echo "Unmounted correctly"
+								echo "Flashing, please wait..."
+								dd if=$IMAGE_PATH/${imageArray[$flash]} | dd of=$FLASH_DEVICE bs=8m
+							else
+								# Bad device number exit
+								clear
+								redColor
+								echo "Please, change the device number"
+								resetColor
+								exit 1
+							fi
 						fi
 					;;
 				* ) clear
@@ -83,22 +112,44 @@ if [ $currentUser == "root" ]; then
 			esac
 		else
 			if [ "$EXTENSION" == "gz" ]; then
-				if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
-					echo "Unmounted correctly"
-					echo "Flashing, please wait..."
-					echo ${imageArray[$flash]}
-					gunzip -c $IMAGE_PATH/${imageArray[$flash]} | pv -s 7g | dd of=$FLASH_DEVICE bs=8m
+				# Check if 'pv' command exists
+				if [ -z $EXISTSPV ]; then
+					if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
+						# Flash with progress bar
+						echo "Unmounted correctly"
+						echo "Flashing, please wait..."
+						echo ${imageArray[$flash]}
+						gunzip -c $IMAGE_PATH/${imageArray[$flash]} | pv -s 7g | dd of=$FLASH_DEVICE bs=8m
+					else
+						# Bad device number exit
+						clear
+						redColor
+						echo "Please, change the device number"
+						resetColor
+						exit 1
+					fi
 				else
-					clear
-					redColor
-					echo "Please, change the device number"
-					resetColor
+					# Flash without progress bar
+					if [ "`diskutil unmountDisk $DEVICE 2>/dev/null`" ]; then
+						echo "Unmounted correctly"
+						echo "Flashing, please wait..."
+						echo ${imageArray[$flash]}
+						gunzip -c $IMAGE_PATH/${imageArray[$flash]} | dd of=$FLASH_DEVICE bs=8m
+					else
+						# Bad device number exit
+						clear
+						redColor
+						echo "Please, change the device number"
+						resetColor
+						exit 1
+					fi
 				fi
 			fi
 		fi
 	fi
 	exit 0
 else
+	# Bad user
 	clear
 	redColor
 	echo "\"sudo\" is needed..."
